Разделение main.js
