+++ js/__mocks__/vars.js
@@ -1,0 +1,36 @@
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const NAMES = [
+  'Артём',
+  'Лёха',
+  'Саня',
+  'Маслёнок',
+  'Николай',
+  'Константин Петрович',
+  'Мигель Иванович',
+  'Шанхай Петрович',
+  'Керосин',
+  'Клаус',
+];
+
+const DESCRIPTIONS = [
+  'Достаточной близости радиозвезд их расстояний и была выдвинута гипотеза.',
+  'Отелло рассвирипело и нижегородские собака.',
+  'Клетке сидит мой пернатый друг на камешке, а в любви.',
+  'Горницу вошел негр, румяный с изображением натюрморта как веник форточку ворвался.',
+  'Ольгу на анне карениной трактор мчался.',
+  'Луч солнца и четырёх пар.',
+  'Hе многие люди могли бы так сделать!',
+  'Ходила на меня напала мысль слова дура боец вспомнил.',
+  'Мушкетеры не слышала от страха невиданное зрелище была маша ногами.',
+  'Французские слова, кроме вороны тело млекопитающего состоит из под пальмой.',
+];
+
+export {COMMENTS, NAMES, DESCRIPTIONS};
